+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 ### Internals
 * Migrated to the final `std::filesystem` implementation on Windows from the experimental one.
+* Exception class InvalidKey is replaced with KeyNotFound, KeyAlreadyUsed, ColumnNotFound and ColumnAlreadyExists
 
 ----------------------------------------------